+++ prog/weaveutil/cni.go
@@ -6,11 +6,11 @@ import (
 	cni "github.com/containernetworking/cni/pkg/skel"
 	"github.com/containernetworking/cni/pkg/version"
 
-	bv "github.com/containernetworking/plugins/pkg/utils/buildversion"
 	weaveapi "github.com/AVENTER-UG/weave/api"
 	"github.com/AVENTER-UG/weave/common"
 	ipamplugin "github.com/AVENTER-UG/weave/plugin/ipam"
 	netplugin "github.com/AVENTER-UG/weave/plugin/net"
+	bv "github.com/containernetworking/plugins/pkg/utils/buildversion"
 )
 
 func cniNet(args []string) error {