+++ prog/weavewait/check_network.go
@@ -1,3 +1,4 @@
+//go:build !iface && !mcast
 // +build !iface,!mcast
 
 package main