+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Master
+
+- ADD: [client] add api call to pull client information from CRM.
+- ADD: include go libs from weaveworks common.
+
 ## Release 2.11.0
 
 - ADD: sbom under reweave/scans